# FrontEnd/src/app/OpenAPI_Product.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "description": "This is a sample Spring Boot RESTful service using OpenAPI 3.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8051",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/products": {
      "post": {
        "tags": ["product-controller"],
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/purchase": {
      "post": {
        "tags": ["product-controller"],
        "operationId": "purchaseProducts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ProductPurchaseRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductPurchaseResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/viewProducts": {
      "get": {
        "tags": ["productview-controller"],
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/viewProducts/{product-id}": {
      "get": {
        "tags": ["productview-controller"],
        "operationId": "findById",
        "parameters": [
          {
            "name": "product-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductRequest": {
        "required": [
          "availableQuantity",
          "categoryId",
          "description",
          "naame",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "naame": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "availableQuantity": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProductPurchaseRequest": {
        "required": ["productId", "quantity"],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ProductPurchaseResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "naame": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "availableQuantity": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "type": "string"
          },
          "categoryDescription": {
            "type": "string"
          }
        }
      }
    }
  }
}